// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/reportes/filtro-autoridades-revocadas.ts
export interface FiltroAutoridadesRevocadas {
    esquema?:string;
    usuario: string;
    idProceso: number;
    idEleccion: number;
    idCentroComputo: number;
    idAmbitoElectoral: number;    
    proceso: string;
    eleccion: string;
    centroComputo: string;
    odpe: string;
    idCargo: number;
}